% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acf_to_sigma}
\alias{acf_to_sigma}
\title{Compute Sigma matrix given an acf vector}
\usage{
acf_to_sigma(acf)
}
\arguments{
\item{acf}{A \code{vector} specifying the autocovariance vector.}
}
\description{
Compute Sigma matrix given an acf vector
}
