% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{autocovariance_to_wv}
\alias{autocovariance_to_wv}
\title{Compute the theoretical WV given the autocovariance vector}
\usage{
autocovariance_to_wv(acf, tau)
}
\arguments{
\item{acf}{A \code{vector} specifying the autocovariance vector}

\item{tau}{A \code{vector} of Wavelet Variance scales}
}
\description{
Compute the theoretical WV given the autocovariance vector
}
