% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{cola}
\alias{cola}
\title{GNSS time series for PBO station COLA}
\format{
A \code{gnssts} object of the East position (dE) of the COLA station
}
\source{
\url{https://data.unavco.org/archive/gnss/products/position/COLA/COLA.pbo.igs14.pos}
}
\usage{
cola
}
\description{
Data from station COLA of the Plate Boundary Observatory
}
\keyword{datasets}
