% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fit.R
\name{compare_fits}
\alias{compare_fits}
\title{Compare graphically two \code{gnsstsmodel} objects.}
\usage{
compare_fits(fit_1, fit_2, main = NULL, y_unit = "mm", x_unit = "days")
}
\arguments{
\item{fit_1}{A \code{gnsstsmodel} object.}

\item{fit_2}{A \code{gnsstsmodel} object.}

\item{main}{A \code{string} specifying the plot title.}

\item{y_unit}{A \code{string} specifying the y axis label.}

\item{x_unit}{A \code{string} specifying the x axis label.}
}
\description{
Compare graphically two \code{gnsstsmodel} objects.
}
