% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fgn.R
\name{fgn_autocovariance}
\alias{fgn_autocovariance}
\title{FGN autocovariance}
\usage{
fgn_autocovariance(theta, n)
}
\arguments{
\item{theta}{A \code{vector} of length 2 specifying the parameters of the FGN}

\item{n}{A \code{scalar} specifying the length of the signal}
}
\description{
FGN autocovariance
}
