% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBO.R
\name{PBO_get_offsets}
\alias{PBO_get_offsets}
\title{Extract offsets for a PBO station}
\usage{
PBO_get_offsets(station_name)
}
\arguments{
\item{station_name}{A \code{string} specifying the PBO station name.}
}
\value{
A \code{vector} specifying the offsets of a PBO station.
}
\description{
Extract offsets for a PBO station
}
\examples{
\dontrun{
pbo_cola_offsets = PBO_get_offsets(station_name = "COLA")
pbo_cola_offsets
}
}
