% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{powerlaw_autocovariance}
\alias{powerlaw_autocovariance}
\title{Compute the power law autocovariance vector}
\usage{
powerlaw_autocovariance(theta, n)
}
\arguments{
\item{theta}{A \code{vector} of length 2 specifying the parameters of the powerlaw process}

\item{n}{A \code{scalar} specifying the length of the signal}
}
\description{
Compute the power law autocovariance vector
}
