% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnssts.R
\name{read.gnssts}
\alias{read.gnssts}
\title{Read a gnssts object}
\usage{
read.gnssts(filename, format = "mom")
}
\arguments{
\item{filename}{A \code{string} specifying the name of the file to read.}

\item{format}{A \code{string} specifying the format of the file to read.}
}
\value{
Return a \code{gnssts} object.
}
\description{
Read a gnssts object
}
