% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector.R
\name{remove_outliers_hector}
\alias{remove_outliers_hector}
\title{Remove outliers from a gnssts object using Hector}
\usage{
remove_outliers_hector(x, n_seasonal, IQ_factor = 3, cleanup = TRUE)
}
\arguments{
\item{x}{A \code{gnssts} object}

\item{n_seasonal}{An \code{integer} specifying the number of seasonal component in the time series.}

\item{IQ_factor}{the \code{IQ_factor} parameter in hector removeoutliers.ctl}

\item{cleanup}{An \code{boolean} specifying if temporary files should be cleaned.}
}
\value{
A \code{gnssts} object.
}
\description{
Remove outliers from a gnssts object using Hector
}
