% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnssts.R
\name{create.gnssts}
\alias{create.gnssts}
\title{Create a gnssts object}
\usage{
create.gnssts(t, y, jumps = NULL, sampling_period = 1)
}
\arguments{
\item{t}{A \code{vector} specifying the time of each observation of the time series.}

\item{y}{A \code{vector} specifying the values of each observation of the time series.}

\item{jumps}{A \code{vector} specifying the time values for which there is a jump.}

\item{sampling_period}{An \code{integer} specifying the sampling period.}
}
\value{
A \code{gnssts} object.
}
\description{
Create a gnssts object
}
