% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gnsstsmodel}
\alias{print.gnsstsmodel}
\title{Print method for a \code{gnsstsmodel} object.}
\usage{
\method{print}{gnsstsmodel}(x, ...)
}
\arguments{
\item{x}{A \code{gnsstsmodel} object.}

\item{...}{Additional graphical parameters.}
}
\description{
Print method for a \code{gnsstsmodel} object.
}
