% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnssts.R
\name{write.gnssts}
\alias{write.gnssts}
\title{Write a gnssts object}
\usage{
write.gnssts(x, filename, format = "mom")
}
\arguments{
\item{x}{A \code{R} object to save as a \code{gnssts} object.}

\item{filename}{A \code{string} specifying the name of the file to write.}

\item{format}{A \code{string} specifying the format of the file to write.}
}
\value{
No return value. Write a \code{gnssts} object in a .mom file by default.
}
\description{
Write a gnssts object
}
